app/requirements.txt
requests~=2.28.2
xmltodict~=0.13.0
kafka-python
cassandra-driver~=3.26.0
Flask~=2.2.2
feedparser~=6.0.10